Implement navigateTo in ExtNavState
