+++ katalog/src/main/kotlin/jp/co/cyberagent/katalog/compose/KatalogViewModel.kt
@@ -6,6 +6,7 @@ import jp.co.cyberagent.katalog.compose.navigation.DiscoveryDestination
 import jp.co.cyberagent.katalog.compose.navigation.ExtNavStateImpl
 import jp.co.cyberagent.katalog.compose.navigation.MainDestination
 import jp.co.cyberagent.katalog.compose.navigation.NavController
+import jp.co.cyberagent.katalog.compose.navigation.navigateTo
 import jp.co.cyberagent.katalog.domain.CatalogItem
 import jp.co.cyberagent.katalog.domain.Katalog
 import jp.co.cyberagent.katalog.domain.KatalogContainer
@@ -34,7 +35,10 @@ internal class KatalogViewModel(
     val errorMessage: StateFlow<String?> = _errorMessage
 
     val navController: NavController<MainDestination> = NavController(initialDestination)
-    private val extNavState: ExtNavState = ExtNavStateImpl(navController)
+    private val extNavState: ExtNavState = ExtNavStateImpl(
+        navController = navController,
+        katalog = katalog
+    )
 
     init {
         viewModelScope.launch(Dispatchers.Default) {
@@ -50,25 +54,6 @@ internal class KatalogViewModel(
     }
 
     fun handleClick(item: CatalogItem) {
-        when (item) {
-            is CatalogItem.Group -> {
-                val currentDestination = navController.current.destination
-                val nextChildDestination = DiscoveryDestination.Group(item)
-                if (currentDestination is MainDestination.Discovery) {
-                    currentDestination.childNavController.push(nextChildDestination)
-                } else {
-                    val nextDestination = MainDestination.Discovery(
-                        childNavController = NavController(nextChildDestination)
-                    )
-                    navController.push(nextDestination)
-                }
-            }
-            is CatalogItem.Component -> {
-                val nextDestination = MainDestination.Preview(
-                    component = item
-                )
-                navController.push(nextDestination)
-            }
-        }
+        navController.navigateTo(item)
     }
 }
